Add files via upload
Plots implemented in my 2025 Lunar and Planetary Science Conference Abstract submission: A Cross-Laboratory Comparison of Titan Haze Analogs: Mechanical Properties. They display the preliminary results for the Young's modulus (GPa) and hardness (GPa) 5% CH4/N2 NASA ARC and JHU plasma tholins and are compared to Yu et al. 2018/2017. The substrates of each sample is also included for comparison and efforts to answer the question of substrate choice influencing the mechanical properties of tholins.